change how links are checked daily/hourly
